requirements.txt
torch
torchvision
transformers
mujoco
gymnasium
numpy
opencv-python
matplotlib
pandas
lerobot
pyyaml
pillow
